requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
python-jose[cryptography]
bcrypt
python-multipart
pydantic[email]
